refactor: Edit Token HttpClient And Import MainActivity Path
