+++ app/src/main/java/com/b1nd/alimo/data/remote/service/ImageService.kt
@@ -3,6 +3,7 @@ package com.b1nd.alimo.data.remote.service
 import com.b1nd.alimo.data.Env.testToken
 import com.b1nd.alimo.data.remote.response.BaseResponse
 import com.b1nd.alimo.data.remote.response.Image.ImageResponse
+import com.b1nd.alimo.di.AppHttpClient
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
@@ -12,7 +13,7 @@ import io.ktor.client.request.parameter
 import javax.inject.Inject
 
 class ImageService @Inject constructor(
-    private val httpClient: HttpClient
+    @AppHttpClient private val httpClient: HttpClient
 ) {
 
     suspend fun getNotificationImage(
