+++ app/src/main/java/com/b1nd/alimo/data/remote/service/PostService.kt
@@ -4,6 +4,7 @@ import com.b1nd.alimo.data.Env
 import com.b1nd.alimo.data.remote.request.detail.DetailEmojiRequest
 import com.b1nd.alimo.data.remote.response.BaseResponse
 import com.b1nd.alimo.data.remote.response.notification.NotificationResponse
+import com.b1nd.alimo.di.AppHttpClient
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
@@ -16,7 +17,7 @@ import io.ktor.client.request.setBody
 import javax.inject.Inject
 
 class PostService @Inject constructor(
-    private val httpClient: HttpClient
+    @AppHttpClient private val httpClient: HttpClient
 ) {
     suspend fun getBookmark(
         page: Int,
